Refactored LongTracking code to separate quantile tracking from basic tracking
